+++ yakssok/src/main/java/server/yakssok/domain/auth/presentation/dto/request/JoinRequest.java
@@ -5,11 +5,11 @@ import jakarta.validation.constraints.NotNull;
 import server.yakssok.domain.user.domain.entity.User;
 
 public record JoinRequest(
-	@Schema(description = "카카오 Access Token/ 애플 authorization code", example = "1234567890abcdef")
+	@Schema(description = "카카오 Access Token/ 애플 idToken", example = "1234567890abcdef")
 	@NotNull
 	String socialAuthorizationCode,
 
-	@Schema(description = "소셜 타입(apple/kakao)", example = "apple")
+	@Schema(description = "소셜 타입(apple/kakao)", example = "kakao")
 	@NotNull
 	String socialType,
