+++ yakssok/src/main/java/server/yakssok/global/infra/oauth/SocialUserResponse.java
@@ -1,4 +1,4 @@
-package server.yakssok.domain.auth.application.client;
+package server.yakssok.global.infra.oauth;
 
 public interface SocialUserResponse {
 	String getId();
