+++ yakssok/src/main/java/server/yakssok/global/infra/oauth/kakao/KakaoUserAccount.java
@@ -1,4 +1,4 @@
-package server.yakssok.domain.auth.application.client.kakao;
+package server.yakssok.global.infra.oauth.kakao;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
